Refactor out reset-password
