+++ auth-service/routes/index.ts
@@ -5,6 +5,7 @@ import { logoutHandler } from "./auth-logout.ts";
 import { registerHandler } from "./auth-register.ts";
 import { verifyHandler } from "./auth-verify.ts";
 import { forgotPasswordHandler } from "./auth-forgot-password.ts";
+import { resetPasswordHandler } from "./auth-reset-password.ts";
 
 const router = express.Router();
 
@@ -13,5 +14,6 @@ router.use("/auth/register", registerHandler);
 router.use("/auth/logout", logoutHandler);
 router.use("/auth/verify", verifyHandler);
 router.use("/auth/forgot-password", forgotPasswordHandler);
+router.use("/auth/reset-password", resetPasswordHandler);
 
 export { router as authRouter };
